refactor: expose more utilities
