+++ xtask/rules_check/Cargo.toml
@@ -9,12 +9,10 @@ publish     = false
 anyhow                = { workspace = true }
 biome_analyze         = { workspace = true }
 biome_configuration   = { workspace = true }
-biome_console         = { workspace = true }
 biome_css_analyze     = { workspace = true }
 biome_css_parser      = { workspace = true }
 biome_css_semantic    = { workspace = true }
 biome_css_syntax      = { workspace = true }
-biome_deserialize     = { workspace = true }
 biome_diagnostics     = { workspace = true }
 biome_fs              = { workspace = true }
 biome_graphql_analyze = { workspace = true }
@@ -27,7 +25,6 @@ biome_js_analyze      = { workspace = true }
 biome_js_parser       = { workspace = true }
 biome_js_syntax       = { workspace = true }
 biome_json_analyze    = { workspace = true }
-biome_json_factory    = { workspace = true }
 biome_json_parser     = { workspace = true }
 biome_json_syntax     = { workspace = true }
 biome_languages       = { workspace = true, features = ["lang_css", "lang_graphql", "lang_js", "lang_json"] }
